Moved DiscordManager to a separate namespace
