+++ src/Osu.Music.Services/ServicesModule.cs
@@ -4,6 +4,7 @@ using Osu.Music.Services.Hotkeys;
 using Osu.Music.Services.Interfaces;
 using Osu.Music.Services.IO;
 using Osu.Music.Services.Localization;
+using Osu.Music.Services.Social;
 using Prism.Ioc;
 using Prism.Modularity;
